rename common files for Signer, Verifier, and JWT test
