+++ src/SignerAlg/common_SignerAlg.ts
