+++ src/VerifierAlg/ES256VerifierAlg.ts
@@ -4,7 +4,7 @@ import type { VerificationMethod } from 'did-resolver'
 import { bytesToHex, EcdsaSignature } from '../util'
 import { verifyBlockchainAccountId } from '../blockchains'
 
-import * as common from './common'
+import * as common_VerifierAlg from './common_VerifierAlg'
 
 const secp256r1 = new EC('p256')
 
@@ -20,7 +20,7 @@ export function verifyES256(
   authenticators: VerificationMethod[]
 ): VerificationMethod {
   const hash: Uint8Array = sha256(data)
-  const sigObj: EcdsaSignature = common.toSignatureObject(signature)
+  const sigObj: EcdsaSignature = common_VerifierAlg.toSignatureObject(signature)
   const fullPublicKeys = authenticators.filter(({ ethereumAddress, blockchainAccountId }) => {
     return typeof ethereumAddress === 'undefined' && typeof blockchainAccountId === 'undefined'
   })
@@ -30,7 +30,7 @@ export function verifyES256(
 
   let signer: VerificationMethod | undefined = fullPublicKeys.find((pk: VerificationMethod) => {
     try {
-      const pubBytes = common.extractPublicKeyBytes(pk)
+      const pubBytes = common_VerifierAlg.extractPublicKeyBytes(pk)
       return secp256r1.keyFromPublic(pubBytes).verify(hash, <SignatureInput>sigObj)
     } catch (err) {
       return false
@@ -52,9 +52,9 @@ export function verifyRecoverableES256(
 ): VerificationMethod {
   let signatures: EcdsaSignature[]
   if (signature.length > 86) {
-    signatures = [common.toSignatureObject(signature, true)]
+    signatures = [common_VerifierAlg.toSignatureObject(signature, true)]
   } else {
-    const so = common.toSignatureObject(signature, false)
+    const so = common_VerifierAlg.toSignatureObject(signature, false)
     signatures = [
       { ...so, recoveryParam: 0 },
       { ...so, recoveryParam: 1 },
@@ -70,7 +70,7 @@ export function verifyRecoverableES256(
     const recoveredAddress: string = toEthereumAddress(recoveredPublicKeyHex)
 
     const signer: VerificationMethod | undefined = authenticators.find((pk: VerificationMethod) => {
-      const keyHex = bytesToHex(common.extractPublicKeyBytes(pk))
+      const keyHex = bytesToHex(common_VerifierAlg.extractPublicKeyBytes(pk))
       return (
         keyHex === recoveredPublicKeyHex ||
         keyHex === recoveredCompressedPublicKeyHex ||
