+++ src/__tests__/JWT/JWT.ES256KSigner.test.ts
@@ -15,10 +15,10 @@ import {
 import { EdDSASigner } from '../../signers/EdDSASigner'
 import { ES256KSigner } from '../../signers/ES256KSigner'
 import { bytesToBase64url, decodeBase64url } from '../../util'
-import { aud, address, did } from './common/common'
-import { CreatedidDocLegacy } from './common/common'
-import { CreatedidDoc } from './common/common'
-import { CreateauddidDoc } from './common/common'
+import { aud, address, did } from './common_Signer_test/common_Signer_test'
+import { CreatedidDocLegacy } from './common_Signer_test/common_Signer_test'
+import { CreatedidDoc } from './common_Signer_test/common_Signer_test'
+import { CreateauddidDoc } from './common_Signer_test/common_Signer_test'
 
 const NOW = 1485321133
 MockDate.set(NOW * 1000 + 123)
