+++ src/__tests__/JWT/common_Signer_test/__tests__/common_Signer_test.test.ts
@@ -1,9 +1,9 @@
-import { greet } from '../common'
-import { CreatedidDocLegacy } from '../common'
-import { CreatedidDoc } from '../common'
-import { CreateauddidDoc } from '../common'
-import { did } from '../common'
-import { aud } from '../common'
+import { greet } from '../common_Signer_test'
+import { CreatedidDocLegacy } from '../common_Signer_test'
+import { CreatedidDoc } from '../common_Signer_test'
+import { CreateauddidDoc } from '../common_Signer_test'
+import { did } from '../common_Signer_test'
+import { aud } from '../common_Signer_test'
 
 const publicKey = '03fdd57adec3d438ea237fe46b33ee1e016eda6b585c3e27ea66686c2ea5358479'
 const keyTypeVerLegacy = 'Secp256k1VerificationKey2018'
